Recharge Bar, and Better UI Hiding
We now hide the spell button with the other buttons when we hit play.

We also have a recharge bar now for spells which will recharge the spell after a set amount of time that we specify. Currently this value is constant we may want to change it to be not constant.